Add files via upload
added HTML File: Basic to show work.
I added multiple scripts to illustrate the answers individually. However I am not sure how to display it all in one scripts page when you view .
I spent a good 6 hours on this so not sure if this is correct. dividually the answer are correct. but just the first script page is showing.